+++ src/main/platforms-config.js
@@ -574,7 +574,19 @@ export const platformsConfig = {
             'adguard_app_mac': true,
         },
     },
-
+    'MAC_V3': {
+        'platform': 'mac',
+        'path': 'mac_v3',
+        'expires': '12 hours',
+        'configuration': {
+            'ignoreRuleHints': false,
+            'replacements': null,
+        },
+        'defines': {
+            'adguard': true,
+            'adguard_app_mac': true,
+        },
+    },
     'ANDROID': {
         'platform': 'android',
         'path': 'android',
@@ -585,6 +597,19 @@ export const platformsConfig = {
         'defines': {
             'adguard': true,
             'adguard_app_android': true,
+        },
+    },
+    'CLI': {
+        'platform': 'cli',
+        'path': 'cli',
+        'expires': '12 hours',
+        'configuration': {
+            'ignoreRuleHints': false,
+            'replacements': null,
+        },
+        'defines': {
+            'adguard': true,
+            'adguard_app_cli': true,
         },
     },
     'EXTENSION_CHROMIUM': {